✨ 抽离公共逻辑plus
